Use a default Signal Desktop directory
It is no longer necessary to specify a location for the Signal Desktop
directory. By default, sigtop will try a system-dependent location. The
new -d option can be used to specify a different location.

See the manual page for details.

This is the second backwards incompatible change.

(This commit does not contain the changes for macOS and Cygwin. These
will be committed on the portable branch.)